UNFINISHED Current Progress
This is not a complete working of this microservice. I unfortunately could not get it done on the 15th but I wanted to turn in what I finished to show that I have started on the microservice. I will be kicking full gear into this on Sunday when it is not 2:40am in the morning.