Working Partially
I'm still working on finishing this for today but all read and create operations are running. I need to get the update and delete functions running as well as find a way to either allow users to enter parameters into the tests.js or document where they need to change the variables.